Separated the tests into separate files for better understanding what is happening
