+++ tests/integration/integration_install_test.go
@@ -5,8 +5,6 @@ import (
 	"crypto/rand"
 	"fmt"
 	"io/ioutil"
-	"net/http"
-	"net/http/httptest"
 	"os"
 	"os/exec"
 	"path/filepath"
@@ -14,15 +12,13 @@ import (
 	"time"
 
 	helmv2 "github.com/fluxcd/helm-controller/api/v2beta1"
+	kustomizev1 "github.com/fluxcd/kustomize-controller/api/v1beta1"
 	"github.com/google/uuid"
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
+	v1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"sigs.k8s.io/controller-runtime/pkg/client"
-
-	kustomizev1 "github.com/fluxcd/kustomize-controller/api/v1beta1"
-	profilesv1 "github.com/weaveworks/profiles/api/v1alpha1"
-	v1 "k8s.io/api/core/v1"
 )
 
 const (
@@ -30,328 +26,7 @@ const (
 	pctlPrivateProfilesRepositoryName = "git@github.com:weaveworks/profiles-examples-private.git"
 )
 
-var _ = Describe("PCTL", func() {
-	Context("search", func() {
-		It("returns the matching profiles", func() {
-			cmd := exec.Command(binaryPath, "search", "nginx")
-			session, err := cmd.CombinedOutput()
-			Expect(err).ToNot(HaveOccurred())
-			expected := "CATALOG/PROFILE               	VERSION	DESCRIPTION                     \n" +
-				"nginx-catalog/weaveworks-nginx	v0.1.0 	This installs nginx.           \t\n" +
-				"nginx-catalog/weaveworks-nginx	v0.1.1 	This installs nginx.           \t\n" +
-				"nginx-catalog/bitnami-nginx   	v0.0.1 	This installs nginx.           \t\n" +
-				"nginx-catalog/nginx           	v2.0.1 	This installs nginx.           \t\n" +
-				"nginx-catalog/some-other-nginx	       	This installs some other nginx.\t\n\n"
-			Expect(string(session)).To(ContainSubstring(expected))
-		})
-
-		When("-o is set to json", func() {
-			It("returns the matching profiles in json", func() {
-				cmd := exec.Command(binaryPath, "search", "-o", "json", "nginx")
-				session, err := cmd.CombinedOutput()
-				Expect(err).ToNot(HaveOccurred())
-				Expect(string(session)).To(ContainSubstring(`{
-    "tag": "weaveworks-nginx/v0.1.0",
-    "catalogSource": "nginx-catalog",
-    "url": "https://github.com/weaveworks/profiles-examples",
-    "name": "weaveworks-nginx",
-    "description": "This installs nginx.",
-    "maintainer": "weaveworks (https://github.com/weaveworks/profiles)",
-    "prerequisites": [
-      "Kubernetes 1.18+"
-    ]
-  },
-  {
-    "tag": "weaveworks-nginx/v0.1.1",
-    "catalogSource": "nginx-catalog",
-    "url": "https://github.com/weaveworks/profiles-examples",
-    "name": "weaveworks-nginx",
-    "description": "This installs nginx.",
-    "maintainer": "weaveworks (https://github.com/weaveworks/profiles)",
-    "prerequisites": [
-      "Kubernetes 1.18+"
-    ]
-  },
-  {
-    "tag": "bitnami-nginx/v0.0.1",
-    "catalogSource": "nginx-catalog",
-    "url": "https://github.com/weaveworks/profiles-examples",
-    "name": "bitnami-nginx",
-    "description": "This installs nginx.",
-    "maintainer": "weaveworks (https://github.com/weaveworks/profiles)",
-    "prerequisites": [
-      "Kubernetes 1.18+"
-    ]
-  },
-  {
-    "tag": "v2.0.1",
-    "catalogSource": "nginx-catalog",
-    "url": "https://github.com/weaveworks/nginx-profile",
-    "name": "nginx",
-    "description": "This installs nginx.",
-    "maintainer": "weaveworks (https://github.com/weaveworks/profiles)",
-    "prerequisites": [
-      "Kubernetes 1.18+"
-    ]
-  },
-  {
-    "catalogSource": "nginx-catalog",
-    "name": "some-other-nginx",
-    "description": "This installs some other nginx."
-  }`))
-			})
-		})
-
-		When("kubeconfig is incorrectly set", func() {
-			It("returns a useful error", func() {
-				cmd := exec.Command(binaryPath, "--kubeconfig=/non-existing/path/kubeconfig", "search", "nginx")
-				session, err := cmd.CombinedOutput()
-				Expect(err).To(HaveOccurred())
-				Expect(string(session)).To(ContainSubstring("failed to create config from kubeconfig path"))
-			})
-		})
-
-		When("a search string is not provided", func() {
-			It("returns a useful error", func() {
-				cmd := exec.Command(binaryPath, "search")
-				session, err := cmd.CombinedOutput()
-				Expect(err).To(HaveOccurred())
-				Expect(string(session)).To(ContainSubstring("argument must be provided"))
-			})
-		})
-
-		When("a search string is provided with all option", func() {
-			It("returns a useful error", func() {
-				cmd := exec.Command(binaryPath, "search", "-all", "nginx")
-				session, err := cmd.CombinedOutput()
-				Expect(err).To(HaveOccurred())
-				Expect(string(session)).To(ContainSubstring("argument must not be provided"))
-			})
-		})
-
-		It("returns all the profiles with search all option", func() {
-			cmd := exec.Command(binaryPath, "search", "--all")
-			session, err := cmd.CombinedOutput()
-			Expect(err).ToNot(HaveOccurred())
-			expected := "CATALOG/PROFILE               	VERSION	DESCRIPTION                     \n" +
-				"nginx-catalog/weaveworks-nginx	v0.1.0 	This installs nginx.           \t\n" +
-				"nginx-catalog/weaveworks-nginx	v0.1.1 	This installs nginx.           \t\n" +
-				"nginx-catalog/bitnami-nginx   	v0.0.1 	This installs nginx.           \t\n" +
-				"nginx-catalog/nginx           	v2.0.1 	This installs nginx.           \t\n" +
-				"nginx-catalog/some-other-nginx	       	This installs some other nginx.\t\n\n"
-			Expect(string(session)).To(ContainSubstring(expected))
-		})
-
-		When("-o is set to json with search all", func() {
-			It("returns the matching profiles in json", func() {
-				cmd := exec.Command(binaryPath, "search", "-a", "-o", "json")
-				session, err := cmd.CombinedOutput()
-				Expect(err).ToNot(HaveOccurred())
-				Expect(string(session)).To(ContainSubstring(`{
-    "tag": "weaveworks-nginx/v0.1.0",
-    "catalogSource": "nginx-catalog",
-    "url": "https://github.com/weaveworks/profiles-examples",
-    "name": "weaveworks-nginx",
-    "description": "This installs nginx.",
-    "maintainer": "weaveworks (https://github.com/weaveworks/profiles)",
-    "prerequisites": [
-      "Kubernetes 1.18+"
-    ]
-  },
-  {
-    "tag": "weaveworks-nginx/v0.1.1",
-    "catalogSource": "nginx-catalog",
-    "url": "https://github.com/weaveworks/profiles-examples",
-    "name": "weaveworks-nginx",
-    "description": "This installs nginx.",
-    "maintainer": "weaveworks (https://github.com/weaveworks/profiles)",
-    "prerequisites": [
-      "Kubernetes 1.18+"
-    ]
-  },
-  {
-    "tag": "bitnami-nginx/v0.0.1",
-    "catalogSource": "nginx-catalog",
-    "url": "https://github.com/weaveworks/profiles-examples",
-    "name": "bitnami-nginx",
-    "description": "This installs nginx.",
-    "maintainer": "weaveworks (https://github.com/weaveworks/profiles)",
-    "prerequisites": [
-      "Kubernetes 1.18+"
-    ]
-  },
-  {
-    "tag": "v2.0.1",
-    "catalogSource": "nginx-catalog",
-    "url": "https://github.com/weaveworks/nginx-profile",
-    "name": "nginx",
-    "description": "This installs nginx.",
-    "maintainer": "weaveworks (https://github.com/weaveworks/profiles)",
-    "prerequisites": [
-      "Kubernetes 1.18+"
-    ]
-  },
-  {
-    "catalogSource": "nginx-catalog",
-    "name": "some-other-nginx",
-    "description": "This installs some other nginx."
-  }`))
-			})
-		})
-	})
-
-	Context("show", func() {
-		It("returns information about the given profile", func() {
-			cmd := exec.Command(binaryPath, "show", "nginx-catalog/weaveworks-nginx")
-			session, err := cmd.CombinedOutput()
-			Expect(err).ToNot(HaveOccurred())
-			Expect(string(session)).To(ContainSubstring("Catalog      \tnginx-catalog                                      \t\n" +
-				"Name         \tweaveworks-nginx                                   \t\n" +
-				"Version      \tv0.1.0                                             \t\n" +
-				"Description  \tThis installs nginx.                               \t\n" +
-				"URL          \thttps://github.com/weaveworks/profiles-examples    \t\n" +
-				"Maintainer   \tweaveworks (https://github.com/weaveworks/profiles)\t\n" +
-				"Prerequisites\tKubernetes 1.18+                                   \t\n"))
-		})
-
-		When("version is used in the catalog", func() {
-			It("shows the right profile", func() {
-				cmd := exec.Command(binaryPath, "show", "nginx-catalog/weaveworks-nginx/v0.1.0")
-				output, err := cmd.CombinedOutput()
-				Expect(err).ToNot(HaveOccurred())
-				Expect(string(output)).To(ContainSubstring("Catalog      \tnginx-catalog                                      \t\n" +
-					"Name         \tweaveworks-nginx                                   \t\n" +
-					"Version      \tv0.1.0                                             \t\n" +
-					"Description  \tThis installs nginx.                               \t\n" +
-					"URL          \thttps://github.com/weaveworks/profiles-examples    \t\n" +
-					"Maintainer   \tweaveworks (https://github.com/weaveworks/profiles)\t\n" +
-					"Prerequisites\tKubernetes 1.18+                                   \t\n"))
-			})
-		})
-
-		When("-o is set to json", func() {
-			It("returns the profile info in json", func() {
-				cmd := exec.Command(binaryPath, "show", "-o", "json", "nginx-catalog/weaveworks-nginx")
-				session, err := cmd.CombinedOutput()
-				Expect(err).ToNot(HaveOccurred())
-				Expect(string(session)).To(ContainSubstring(`{
-  "tag": "weaveworks-nginx/v0.1.0",
-  "catalogSource": "nginx-catalog",
-  "url": "https://github.com/weaveworks/profiles-examples",
-  "name": "weaveworks-nginx",
-  "description": "This installs nginx.",
-  "maintainer": "weaveworks (https://github.com/weaveworks/profiles)",
-  "prerequisites": [
-    "Kubernetes 1.18+"
-  ]
-}`))
-			})
-		})
-
-		When("a name argument is not provided", func() {
-			It("returns a useful error", func() {
-				cmd := exec.Command(binaryPath, "show")
-				session, err := cmd.CombinedOutput()
-				Expect(err).To(HaveOccurred())
-				Expect(string(session)).To(ContainSubstring("argument must be provided"))
-			})
-		})
-	})
-
-	Context("list", func() {
-		var (
-			namespace        = "default"
-			installationName = "long-name-to-ensure-padding"
-			ctx              = context.TODO()
-			pSub             profilesv1.ProfileInstallation
-		)
-
-		BeforeEach(func() {
-			profileURL := "https://github.com/weaveworks/profiles-examples"
-			pSub = profilesv1.ProfileInstallation{
-				TypeMeta: metav1.TypeMeta{
-					Kind:       "ProfileInstallation",
-					APIVersion: "profileinstallations.weave.works/v1alpha1",
-				},
-				ObjectMeta: metav1.ObjectMeta{
-					Name:      installationName,
-					Namespace: namespace,
-				},
-				Spec: profilesv1.ProfileInstallationSpec{
-					Source: &profilesv1.Source{
-						URL: profileURL,
-						Tag: "weaveworks-nginx/v0.1.0",
-					},
-					Catalog: &profilesv1.Catalog{
-						Catalog: "nginx-catalog",
-						Profile: "weaveworks-nginx",
-						Version: "v0.1.0",
-					},
-				},
-			}
-			Expect(kClient.Create(ctx, &pSub)).Should(Succeed())
-		})
-
-		AfterEach(func() {
-			Expect(kClient.Delete(ctx, &pSub)).Should(Succeed())
-		})
-
-		It("returns the installations", func() {
-			listCmd := func() []string {
-				cmd := exec.Command(binaryPath, "list")
-				session, err := cmd.CombinedOutput()
-				Expect(err).ToNot(HaveOccurred())
-				return strings.Split(string(session), "\n")
-			}
-
-			Eventually(listCmd).Should(ContainElements(
-				"NAMESPACE\tNAME                       \tSOURCE                               \tAVAILABLE UPDATES ",
-				"default  \tlong-name-to-ensure-padding\tnginx-catalog/weaveworks-nginx/v0.1.0\tv0.1.1           \t",
-			))
-		})
-
-		When("there are no available updates", func() {
-			It("returns the installations", func() {
-				profileURL := "https://github.com/weaveworks/profiles-examples"
-				bitnamiSub := profilesv1.ProfileInstallation{
-					TypeMeta: metav1.TypeMeta{
-						Kind:       "ProfileInstallation",
-						APIVersion: "profileinstallations.weave.works/v1alpha1",
-					},
-					ObjectMeta: metav1.ObjectMeta{
-						Name:      "bitnami-profile",
-						Namespace: namespace,
-					},
-					Spec: profilesv1.ProfileInstallationSpec{
-						Source: &profilesv1.Source{
-							URL: profileURL,
-							Tag: "bitnami-nginx/v0.0.1",
-						},
-						Catalog: &profilesv1.Catalog{
-							Catalog: "nginx-catalog",
-							Profile: "bitnami-nginx",
-							Version: "v0.0.1",
-						},
-					},
-				}
-				Expect(kClient.Create(ctx, &bitnamiSub)).Should(Succeed())
-				listCmd := func() []string {
-					cmd := exec.Command(binaryPath, "list")
-					session, err := cmd.CombinedOutput()
-					Expect(err).ToNot(HaveOccurred())
-					return strings.Split(string(session), "\n")
-				}
-
-				Eventually(listCmd).Should(ContainElements(
-					"NAMESPACE\tNAME                       \tSOURCE                               \tAVAILABLE UPDATES ",
-					"default  \tbitnami-profile            \tnginx-catalog/bitnami-nginx/v0.0.1   \t-                \t",
-					"default  \tlong-name-to-ensure-padding\tnginx-catalog/weaveworks-nginx/v0.1.0\tv0.1.1           \t",
-				))
-				Expect(kClient.Delete(ctx, &bitnamiSub)).Should(Succeed())
-			})
-		})
-	})
-
+var _ = Describe("pctl install", func() {
 	Context("install", func() {
 		var (
 			temp      string
@@ -835,7 +510,6 @@ status: {}
 			})
 		})
 	})
-
 	Context("installing from a single-profile repo", func() {
 		var (
 			temp          string
@@ -1095,112 +769,6 @@ status: {}
 					return podList.Items[0].Status.Phase
 				}, 2*time.Minute, 5*time.Second).Should(Equal(v1.PodPhase("Running")))
 				Expect(podList.Items[0].Spec.Containers[0].Image).To(Equal("docker.io/bitnami/nginx:1.21.0-debian-10-r0"))
-			})
-		})
-	})
-
-	Context("prepare", func() {
-		When("dry-run is provided", func() {
-			It("displays the to be applied content", func() {
-				cmd := exec.Command(binaryPath, "prepare", "--dry-run")
-				output, err := cmd.CombinedOutput()
-				Expect(err).ToNot(HaveOccurred())
-				Expect(string(output)).To(ContainSubstring("kind: List"))
-			})
-		})
-		When("baseurl is provided", func() {
-			It("will use that to fetch releases", func() {
-				server := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-					w.WriteHeader(http.StatusTeapot)
-				}))
-				cmd := exec.Command(binaryPath, "prepare", "--baseurl="+server.URL)
-				output, err := cmd.CombinedOutput()
-				Expect(err).To(HaveOccurred())
-				Expect(string(output)).To(ContainSubstring("status: 418 I'm a teapot"))
-			})
-		})
-		When("version is provided", func() {
-			It("will fetch that specific version", func() {
-				// use dry-run here so we don't overwrite the created test cluster resources with old version.
-				cmd := exec.Command(binaryPath, "prepare", "--version=v0.0.1", "--dry-run")
-				output, err := cmd.CombinedOutput()
-				Expect(err).ToNot(HaveOccurred())
-				Expect(string(output)).To(ContainSubstring("kind: List"))
-			})
-		})
-		When("the provided version is missing", func() {
-			It("will put out an understandable error message", func() {
-				cmd := exec.Command(binaryPath, "prepare", "--version=vnope")
-				output, err := cmd.CombinedOutput()
-				Expect(err).To(HaveOccurred())
-				Expect(string(output)).To(ContainSubstring("status: 404 Not Found"))
-			})
-		})
-	})
-
-	Context("docgen", func() {
-		var tmpDir string
-
-		BeforeEach(func() {
-			var err error
-			tmpDir, err = ioutil.TempDir("", "docgen")
-			Expect(err).NotTo(HaveOccurred())
-		})
-
-		AfterEach(func() {
-			_ = os.RemoveAll(tmpDir)
-		})
-
-		It("writes pctl command help to markdown files", func() {
-			cmd := exec.Command(binaryPath, "docgen", "--path", tmpDir)
-			_, err := cmd.CombinedOutput()
-			Expect(err).ToNot(HaveOccurred())
-
-			files, err := ioutil.ReadDir(tmpDir)
-			Expect(err).NotTo(HaveOccurred())
-			Expect(len(files)).To(Equal(5))
-			commands := []string{"install", "prepare", "list", "show", "search"}
-			for _, cmd := range commands {
-				filename := filepath.Join(tmpDir, fmt.Sprintf("pctl-%s-cmd.md", cmd))
-				Expect(filename).To(BeAnExistingFile())
-				contents, err := ioutil.ReadFile(filename)
-				Expect(err).NotTo(HaveOccurred())
-				Expect(string(contents)).To(ContainSubstring(cmd))
-				Expect(string(contents)).To(ContainSubstring("NAME"))
-				Expect(string(contents)).To(ContainSubstring("USAGE"))
-			}
-		})
-
-		When("the provided output directory does not exist", func() {
-			It("creates it", func() {
-				newDir := filepath.Join(tmpDir, "does-not-exist")
-				Expect(newDir).ToNot(BeAnExistingFile())
-
-				cmd := exec.Command(binaryPath, "docgen", "--path", newDir)
-				_, err := cmd.CombinedOutput()
-				Expect(err).ToNot(HaveOccurred())
-				Expect(newDir).To(BeAnExistingFile())
-			})
-
-			When("creating the docs dir fails", func() {
-				It("exits 1", func() {
-					newDir := filepath.Join(tmpDir, "does-not-exist")
-					Expect(os.Chmod(tmpDir, 0600)).To(Succeed())
-
-					cmd := exec.Command(binaryPath, "docgen", "--path", newDir)
-					_, err := cmd.CombinedOutput()
-					Expect(err).To(HaveOccurred())
-				})
-			})
-		})
-
-		When("writing the docs files fails", func() {
-			It("exits 1", func() {
-				Expect(os.Chmod(tmpDir, 0600)).To(Succeed())
-
-				cmd := exec.Command(binaryPath, "docgen", "--path", tmpDir)
-				_, err := cmd.CombinedOutput()
-				Expect(err).To(HaveOccurred())
 			})
 		})
 	})
